+++ src/main/java/mod/tjt01/lapislib/LapisLib.java
@@ -4,6 +4,7 @@ import mod.tjt01.lapislib.client.config.screen.RootConfigScreen;
 import mod.tjt01.lapislib.core.config.LapisLibConfig;
 import mod.tjt01.lapislib.core.network.LapisLibPacketHandler;
 import mod.tjt01.lapislib.registry.loot.GlobalLootModifiers;
+import mod.tjt01.lapislib.registry.loot.LapisLibLootConditions;
 import net.minecraftforge.client.ConfigScreenHandler;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.ModLoadingContext;
@@ -27,6 +28,7 @@ public class LapisLib {
         LapisLibPacketHandler.register();
 
         GlobalLootModifiers.register(bus);
+        LapisLibLootConditions.register(bus);
 
         modLoadingContext.registerExtensionPoint(ConfigScreenHandler.ConfigScreenFactory.class, () ->
                 RootConfigScreen.builder(MODID).getFactory()