+++ src/main/java/mod/tjt01/lapislib/data/loot/modifier/RemoveItemModifier.java
@@ -52,6 +52,6 @@ public class RemoveItemModifier extends LootModifier {
 
     @Override
     public Codec<? extends IGlobalLootModifier> codec() {
-        return null;
+        return CODEC;
     }
 }