+++ src/main/java/mod/tjt01/lapislib/data/loot/modifier/ReplaceItemModifier.java
@@ -58,6 +58,6 @@ public class ReplaceItemModifier extends LootModifier {
 
     @Override
     public Codec<? extends IGlobalLootModifier> codec() {
-        return null;
+        return CODEC;
     }
 }