+++ src/test/java/mod/tjt01/lapislibtest/data/gen/LootModifiers.java
@@ -3,18 +3,19 @@ package mod.tjt01.lapislibtest.data.gen;
 import mod.tjt01.lapislib.data.loot.modifier.AddEntryModifier;
 import mod.tjt01.lapislib.data.loot.modifier.RemoveItemModifier;
 import mod.tjt01.lapislib.data.loot.modifier.ReplaceItemModifier;
+import mod.tjt01.lapislib.data.loot.modifier.condition.BlockTagCondition;
 import mod.tjt01.lapislibtest.LapisLibTest;
 import net.minecraft.data.DataGenerator;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.Ingredient;
-import net.minecraft.world.level.storage.loot.LootPool;
 import net.minecraft.world.level.storage.loot.entries.LootItem;
 import net.minecraft.world.level.storage.loot.functions.LootItemFunction;
 import net.minecraft.world.level.storage.loot.functions.LootingEnchantFunction;
 import net.minecraft.world.level.storage.loot.functions.SetItemCountFunction;
+import net.minecraft.world.level.storage.loot.predicates.ExplosionCondition;
 import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
-import net.minecraft.world.level.storage.loot.providers.number.ConstantValue;
+import net.minecraft.world.level.storage.loot.predicates.LootItemRandomChanceCondition;
 import net.minecraft.world.level.storage.loot.providers.number.UniformGenerator;
 import net.minecraftforge.common.Tags;
 import net.minecraftforge.common.data.GlobalLootModifierProvider;
@@ -54,6 +55,16 @@ public class LootModifiers extends GlobalLootModifierProvider {
                         new LootTableIdCondition.Builder(new ResourceLocation("entities/iron_golem")).build()
                 },
                 Ingredient.of(Tags.Items.INGOTS_IRON)
+        ));
+
+        add("test_block_tag", new AddEntryModifier(
+                new LootItemCondition[] {
+                        LootItemRandomChanceCondition.randomChance(0.25F).build(),
+                        ExplosionCondition.survivesExplosion().build(),
+                        BlockTagCondition.hasTag(Tags.Blocks.ORES_IN_GROUND_STONE).build()
+                },
+                LootItem.lootTableItem(Items.COBBLESTONE).build(),
+                new LootItemFunction[] {}
         ));
     }
 }